requirements.txt
funcsigs==0.4
mock==1.3.0
pbr==1.8.1
requests==2.9.1
setproctitle==1.1.9
six==1.10.0
websocket-client==0.35.0
wheel==0.24.0
